requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
python-dotenv